fix processing + debug

+++ src/openqdc/datasets/multixcqm9.py
@@ -522,5 +522,5 @@ class MultixcQM9(BaseDataset):
         df_xyz = self._read_all_xyzs()
         return [
             {"energies": np.atleast_2d(en), **xyz_dict}
-            for xyz_dict, en in zip(df_xyz.to_dict("records"), df_energies.values)
+            for xyz_dict, en in zip(df_xyz.to_dict("records"), df_energies.values.astype(np.float32))
         ]
